Implement FileDownloader, first attempt
This introduces a new class, FileDownloader, which uses Electron
`DownloadItem` to download files. It is meant to replace the use of
electron-remote for this purpose. However, it doesn't work, because you
can't pass callbacks from renderer processes to the main process and
expect them to run synchronously (like we need in this case).